fix parse_dtype
