+++ tests/dataloaders/test_sequence.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 from pybedtools import Interval
 from kipoi.utils import override_default_kwargs
 from kipoiseq.transforms.functional import one_hot_dna
-from kipoiseq.dataloaders.sequence import IntervalSeqStringDl, IntervalSeqDl, parse_dtype, BedDataset
+from kipoiseq.dataloaders.sequence import IntervalSeqStringDl, IntervalSeqDl, BedDataset
 
 
 @pytest.fixture
@@ -32,15 +32,6 @@ def test_min_props():
         assert all([el in props for el in min_set_props])
 
 
-def test_parse_dtype():
-    dtypes = {'int': int, 'string': str, 'float': float, 'bool': bool}
-    assert all([parse_dtype(dt) == dtypes[dt] for dt in dtypes.keys()])
-    assert all([parse_dtype(dt) == dt for dt in dtypes.values()])
-    with pytest.raises(Exception):
-        parse_dtype("int8")
-    assert parse_dtype(None) is None
-
-
 def test_fasta_based_dataset(intervals_file, fasta_file):
     # just test the functionality
     dl = IntervalSeqStringDl(intervals_file, fasta_file)
@@ -52,7 +43,7 @@ def test_fasta_based_dataset(intervals_file, fasta_file):
     # with pytest.raises(Exception):
     #     dl[0]
 
-    dl = IntervalSeqStringDl(intervals_file, fasta_file, label_dtype="string")
+    dl = IntervalSeqStringDl(intervals_file, fasta_file, label_dtype="str")
     ret_val = dl[0]
     assert isinstance(ret_val['targets'][0], np.str_)
     dl = IntervalSeqStringDl(intervals_file, fasta_file, label_dtype="int")
@@ -72,6 +63,44 @@ def test_seq_dataset(intervals_file, fasta_file):
     assert np.all(ret_val['inputs'] == one_hot_dna("GT"))
     assert isinstance(ret_val["inputs"], np.ndarray)
     assert ret_val["inputs"].shape == (2, 4)
+
+
+@pytest.fixture
+def example_kwargs():
+    return IntervalSeqDl.example_kwargs
+
+
+@pytest.mark.parametrize("alphabet_axis", list(range(0, 4)))
+@pytest.mark.parametrize("dummy_axis", [None] + list(range(0, 4)))
+def test_seq_dataset_reshape(alphabet_axis, dummy_axis, example_kwargs):
+    seq_len, alphabet_len = 3, 4
+
+    kwargs = example_kwargs
+    kwargs['auto_resize_len'] = seq_len
+    kwargs['alphabet_axis'] = alphabet_axis
+    kwargs['dummy_axis'] = dummy_axis
+
+    dummy_axis_int = dummy_axis
+    if dummy_axis is None:
+        dummy_axis_int = -2
+
+    if (alphabet_axis == dummy_axis_int) or (alphabet_axis == -1) or (dummy_axis_int == -1) or \
+            (alphabet_axis >= 3) or (dummy_axis_int >= 3) or ((alphabet_axis >= 2) and (dummy_axis is None)):
+        with pytest.raises(Exception):
+            seq_dataset = IntervalSeqDl(**kwargs)
+        return None
+
+    seq_dataset = IntervalSeqDl(**kwargs)
+
+    # test the single sample works
+    reshaped = seq_dataset[0]['inputs']
+    for i in range(len(reshaped.shape)):
+        if i == dummy_axis:
+            assert reshaped.shape[i] == 1
+        elif i == alphabet_axis:
+            assert reshaped.shape[i] == alphabet_len
+        else:
+            assert reshaped.shape[i] == seq_len
 
 
 # download example files
