+++ tests/test_0_transforms_class.py
@@ -1,44 +1,16 @@
 import pytest
 import numpy as np
 import copy
-from kipoiseq.transforms.functional import resize_interval
-from kipoiseq.transforms.transforms import SplitSplicingSeq, ReorderedOneHot
+from kipoiseq.transforms.transforms import Compose, OneHot, SplitSplicingSeq, ReorderedOneHot
 from kipoiseq.utils import DNA
 from pybedtools import Interval
 
 
 # --------------------------------------------
-
-@pytest.mark.parametrize("anchor", ['start', 'end', 'center'])
-@pytest.mark.parametrize("ilen", [3, 4])
-def test_resize_interval(anchor, ilen):
-    import pybedtools
-    dummy_start, dummy_end = 10, 20
-    dummy_centre = int((dummy_start + dummy_end) / 2)
-
-    dummy_inter = pybedtools.create_interval_from_list(['chr2', dummy_start, dummy_end, 'intname'])
-    ret_inter = resize_interval(dummy_inter, ilen, anchor)
-
-    # the original interval was left intact
-    assert dummy_inter.chrom == 'chr2'
-    assert dummy_inter.start == dummy_start
-    assert dummy_inter.end == dummy_end
-    assert dummy_inter.name == 'intname'
-
-    # metadata kept
-    assert ret_inter.chrom == dummy_inter.chrom
-    assert ret_inter.name == 'intname'
-
-    # desired output width
-    assert ret_inter.length == ilen
-
-    # correct anchor point
-    if anchor == "start":
-        assert ret_inter.start == dummy_start
-    elif anchor == "end":
-        assert ret_inter.end == dummy_end
-    elif anchor == "centre":
-        assert int((ret_inter.start + ret_inter.end) / 2) == dummy_centre
+def test_compose():
+    c = Compose([OneHot()])
+    print(str(c))
+    assert c("ACGT").shape == (4, 4)
 
 
 def test_ReorderedOneHot():
@@ -60,10 +32,10 @@ def test_ReorderedOneHot():
         assert out.shape == tr.get_output_shape(seqlen)
         assert out.shape == result
 
-    with pytest.raises(Exception):
+    with pytest.raises(ValueError):
         ReorderedOneHot(alphabet_axis=1, dummy_axis=1)
 
-    with pytest.raises(Exception):
+    with pytest.raises(ValueError):
         ReorderedOneHot(dummy_axis=1)
 
 
@@ -86,3 +58,7 @@ def test_SplitSplicingSeq():
     assert splited['exon'] == 'GTAGTAGA'
     assert splited['donor'] == 'AGAGT'
     assert splited['intron3prime'] == 'CC'
+
+
+def test_ResizeInterval():
+    pass
